refactor(chat): share one companion-turn pipeline between app and overlay
The app and overlay pages each carried a near-duplicate processCompanionResponse.
The overlay copy was missing the forced-JSON extraction fallback, so weak models
that skip the inline state block silently lost their mood/memory updates there.
Extract the whole turn (parse, extraction fallback, state/mood apply, fact +
turn persistence, stage transition, event check) into processCompanionTurn and
have both pages call it, returning the dialogue, new memory, and any triggered
event so page-specific effects (event modal, photo keepsakes) stay in the pages.
The overlay now gets the extraction fallback too.
